Removed unneeded casts.

+++ ptolemy/data/properties/lattice/typeSystem_C/UChar.java
@@ -57,7 +57,7 @@ public class UChar extends LatticeProperty implements TypeProperty {
    }
 
    public Token getMinValue() {
-       return new LongToken((long)0);
+       return new LongToken(0);
    }
 
    public boolean hasMinMaxValue() {
